+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Allow search by words (establishment name)
 - Allow search by GIAS establishment number
+- Add `New` project table view for your team projects
 
 ## [Release-44][release-44]